:construction: Rework AIG algorithms to their own translation units
